Revert "[PIL-1099] - Page for referral to UK Tax Return"
