+++ app/config/FrontendAppConfig.scala
@@ -35,12 +35,10 @@ class FrontendAppConfig @Inject() (configuration: Configuration, servicesConfig:
   def feedbackUrl(implicit request: RequestHeader): String =
     s"$contactHost/contact/beta-feedback?service=$contactFormServiceIdentifier&backUrl=${SafeRedirectUrl(host + request.uri).encodedUrl}"
 
-  val loginUrl:                   String = configuration.get[String]("urls.login")
-  val loginContinueUrl:           String = configuration.get[String]("urls.loginContinue")
-  val signOutUrl:                 String = configuration.get[String]("urls.signOut")
-  val asaHomePageUrl:             String = configuration.get[String]("urls.asaHomePage")
-  val changeAccountingPeriodUrl:  String = configuration.get[String]("urls.changeAccountingPeriod")
-  val ukTaxReturnSupplierListUrl: String = configuration.get[String]("urls.ukTaxReturnSupplierList")
+  val loginUrl:         String = configuration.get[String]("urls.login")
+  val loginContinueUrl: String = configuration.get[String]("urls.loginContinue")
+  val signOutUrl:       String = configuration.get[String]("urls.signOut")
+  val asaHomePageUrl:   String = configuration.get[String]("urls.asaHomePage")
 
   val pillar2BaseUrl: String = servicesConfig.baseUrl("pillar2")
