+++ app/controllers/actions/DataRetrievalAction.scala
@@ -30,7 +30,7 @@ class DataRetrievalActionImpl @Inject() (
 
   override protected def transform[A](request: IdentifierRequest[A]): Future[OptionalDataRequest[A]] =
     sessionRepository.get(request.userId).map {
-      OptionalDataRequest(request.request, request.userId, _, Some(request.enrolments))
+      OptionalDataRequest(request.request, request.userId, _)
     }
 }
