+++ app/models/subscription/SubscriptionData.scala
@@ -33,10 +33,3 @@ object SubscriptionData {
 
   implicit val format: OFormat[SubscriptionData] = Json.format[SubscriptionData]
 }
-
-final case class SubscriptionSuccess(success: SubscriptionData)
-
-object SubscriptionSuccess {
-
-  implicit val format: OFormat[SubscriptionSuccess] = Json.format[SubscriptionSuccess]
-}
