--- app/utils/Pillar2Reference.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package utils
-
-import uk.gov.hmrc.auth.core.Enrolment
-
-object Pillar2Reference {
-  def getPillar2ID(enrolments: Option[Set[Enrolment]], key: String, identifier: String): Option[String] = enrolments.flatMap { enrolments =>
-    enrolments
-      .find(_.key.equalsIgnoreCase(key))
-      .flatMap(_.identifiers.find(_.key.equalsIgnoreCase(identifier)))
-      .map(_.value)
-  }
-
-}
