+++ app/views/UnauthorisedView.scala.html
@@ -20,7 +20,6 @@
 @this(
         layout: templates.Layout,
         paragraphBody: ParagraphBody,
-        link: Link,
         heading: Heading
 )
 
@@ -35,9 +34,6 @@
 ) {
 
     @heading(messages("unauthorised.heading"))
-    @paragraphBody(messages("unauthorised.guidance", link(
-        text = messages("unauthorised.guidance.link"),
-        call = Call("GET", appConfig.startPagePillar2SubmissionUrl)
-    )))
+    @paragraphBody(messages("unauthorised.guidance"))
 
 }
