--- app/views/components/gds/InactiveStatusBanner.scala.html
@@ -1,34 +1,0 @@
-@*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@this()
-
-@()(implicit messages: Messages)
-
-<div class="govuk-notification-banner" role="region"
-aria-labelledby="govuk-notification-banner-title"
-data-module="govuk-notification-banner">
-    <div class="govuk-notification-banner__header">
-        <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">
-            @messages("uktr.UkTaxReturnStart.inactive.subheading")
-        </h2>
-    </div>
-    <div class="govuk-notification-banner__content">
-        <p class="govuk-notification-banner__heading">
-            @messages("uktr.UkTaxReturnStart.inactive.p1")
-        </p>
-    </div>
-</div>
