--- app/views/uktr/UkTaxReturnStartView.scala.html
@@ -1,51 +1,0 @@
-@*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import config.FrontendAppConfig
-@import uk.gov.hmrc.govukfrontend.views.html.components.HtmlContent
-@import views.html.components.gds._
-@this(
-        layout: templates.Layout,
-        h2: HeadingH2,
-        p: ParagraphBody,
-        heading: Heading,
-        pWithLink: ParagraphMessageWithLink,
-        inactiveStatusBanner: InactiveStatusBanner
-)
-
-@(inactiveStatus: Boolean)(implicit request: Request[_], appConfig: FrontendAppConfig, messages: Messages)
-
-@layout(
-    pageTitle    = titleNoForm(messages("uktr.UkTaxReturnStart.title"))
-) {
-    @if(inactiveStatus) {@inactiveStatusBanner()}
-    @heading(messages("uktr.UkTaxReturnStart.header"), classes = "govuk-heading-l")
-    @p(messages("uktr.UkTaxReturnStart.p1"))
-    @p(messages("uktr.UkTaxReturnStart.p2"))
-    @p(messages("uktr.UkTaxReturnStart.p3"))
-    @h2(messages("uktr.UkTaxReturnStart.subheading"), size = "m")
-    @pWithLink(
-        message1 = Some(messages("uktr.UkTaxReturnStart.p4")),
-        linkMessage = messages("uktr.UkTaxReturnStart.p4.link"),
-        linkUrl = appConfig.changeAccountingPeriodUrl,
-        linkFullStop = true
-    )
-    @p(messages("uktr.UkTaxReturnStart.p5"))
-    @pWithLink(
-        linkMessage = messages("uktr.UkTaxReturnStart.p6.link"),
-        linkUrl = appConfig.ukTaxReturnSupplierListUrl
-    )
-}
