+++ conf/app.routes
@@ -36,4 +36,3 @@ GET         /asa/error/individual                        controllers.AgentContro
 GET         /asa/error/organisation                      controllers.AgentController.onPageLoadOrganisationError
 GET         /asa/home                                    controllers.ASAStubController.onPageLoad
 
-GET         /uk-tax-return                               controllers.uktr.UkTaxReturnStartController.onPageLoad
