+++ conf/application.conf
@@ -104,8 +104,6 @@ urls {
   signOut                    = "http://localhost:9553/bas-gateway/sign-out-without-state"
   startPagePillar2Submission = "https://www.gov.uk/guidance/report-pillar-2-top-up-taxes"
   asaHomePage                = "/report-pillar2-submission-top-up-taxes/asa/home"
-  changeAccountingPeriod     = "http://localhost:10050/report-pillar2-top-up-taxes/manage-account/further-details/change-accounting-period"
-  ukTaxReturnSupplierList    = "https://www.gov.uk/guidance/report-pillar-2-top-up-taxes"
 }
 
 enrolment {
